requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
rank-bm25
pydantic
python-dotenv
pymupdf
httpx
pyjwt
python-telegram-bot==20.8
supervisor
